refactor(transforms): move `ToSample` into shared transforms module
- Move `ToSample` from `src/transforms/vision/ops.rs` to `src/transforms/mod.rs` so it can be reused across modalities.
